fix sendmany
sendmany's 3rd argument is "minimum confs", meaning it will only attempt to spend coins with this many confirmations. The default value of this is 1. This causes problems with mempool spending.

The "start_height" key no longer exists in snapshot jsons, and it was never used in this script to begin with.